+++ mipengine/udfgen/__init__.py
@@ -10,9 +10,9 @@ from mipengine.udfgen.iotypes import state
 from mipengine.udfgen.iotypes import tensor
 from mipengine.udfgen.iotypes import transfer
 from mipengine.udfgen.iotypes import udf_logger
+from mipengine.udfgen.py_udfgenerator import FlowUdfArg
+from mipengine.udfgen.py_udfgenerator import UdfGenerator
 from mipengine.udfgen.smpc import secure_transfer
-from mipengine.udfgen.udfgenerator import FlowUdfArg
-from mipengine.udfgen.udfgenerator import UdfGenerator
 
 __all__ = [
     "literal",
